+++ source/common/sys_info/sys_info_widget_summary.cc
@@ -266,7 +266,7 @@ void SysInfoWidgetSummary::setSystemInfo(const proto::system_info::SystemInfo& s
                     group << mk(tr("Manufacturer"), module.manufacturer());
 
                 if (module.size())
-                    group << mk(tr("Size"), Formatter::sizeToString(static_cast<qint64>(module.size())));
+                    group << mk(tr("Size"), Formatter::sizeToString(module.size()));
 
                 if (module.speed())
                     group << mk(tr("Speed"), tr("%1 MHz").arg(module.speed()));
@@ -322,8 +322,8 @@ void SysInfoWidgetSummary::setSystemInfo(const proto::system_info::SystemInfo& s
             if (drive.total_size() && drive.total_size() != static_cast<quint64>(-1))
             {
                 value = tr("%1 (%2 free)")
-                    .arg(Formatter::sizeToString(static_cast<qint64>(drive.total_size())),
-                         Formatter::sizeToString(static_cast<qint64>(drive.free_size())));
+                    .arg(Formatter::sizeToString(drive.total_size()),
+                         Formatter::sizeToString(drive.free_size()));
             }
 
             items << mk(param, value);
